Breakout Re Usable Functions Into Utils
